created own form_component for membership new_edit
